requirements.txt
importlib-resources==5.4.0
matplotlib==3.5.0
matplotlib-inline==0.1.3
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==8.4.0
scikit-learn==1.0.1
torch==1.10.0+cpu
torchaudio==0.10.0+cpu
torchvision==0.11.1+cpu
tqdm==4.62.3